+++ R/activate.R
@@ -57,7 +57,8 @@ activate.tidync <- function(.data, what, ..., select_var = NULL) {
 
   if (is.numeric(what)) {
     ## this pattern is copied from print
-    ushapes <- dplyr::distinct(.data$grid) %>% 
+    ## remove $variables because it a list 
+    ushapes <- dplyr::distinct(.data$grid %>% dplyr::select(-.data$variables)) %>% 
       dplyr::arrange(desc(nchar(.data$grid)))
     ## otherwise pick the what-th grid
     stopifnot(what >= 1 && what <= nrow(.data$grid))