+++ NEWS.md
@@ -1,5 +1,8 @@
 # dev
 
+* `axis_transforms` is deprecated by `hyper_transforms()` which includes an argument
+`all` to control which set of transforms is returned. 
+
 * Now supports 'NC_CHAR' type, by exploding these into the array size expected. 
 
 * Breaking change: when using `tidync$grid`it's now expected that this must